refactor: move EBookingStatus enum to a seperated file
