+++ API/src/classes/EBookingStatus.js
@@ -1,0 +1,13 @@
+/**
+ * Enum for accommodation amenity types.
+ * @readonly
+ * @enum {string}
+ */
+const EBookingStatus = Object.freeze({
+	BOOKED: "BOOKED",
+	RESERVED: "RESERVED",
+	CHECKED_IN: "CHECKED_IN",
+	COMPLETED: "COMPLETED",
+	CANCELED: "CANCELED",
+});
+export default EBookingStatus;
