+++ API/src/classes/index.js
@@ -2,7 +2,8 @@ export { default as Accommodation } from "./Accommodation.js";
 export { default as AccommodationAmenity, EAccommodationAmenityType } from "./AccommodationAmenity.js";
 export { default as Address } from "./Address.js";
 export { default as Amenity } from "./Amenity.js";
-export { default as Booking, EBookingStatus } from "./Booking.js";
+export { default as Booking } from "./Booking.js";
+export { default as EBookingStatus } from "./EBookingStatus.js";
 export { default as FavouriteList } from "./FavouriteList.js";
 export { default as Image } from "./Image.js";
 export { default as BookingItem } from "./BookingItem.js";
